perf: lazy load dev, dev-exec commands
Non-scientific benchmark says: On my machine, this makes every command
boot 30% faster.
